Refactor dropdown menu component to be a single file
