--- src/components/ui/dropdown-menu/index.ts
@@ -1,10 +1,0 @@
-export * from './dropdown-menu-checkbox-item';
-export * from './dropdown-menu-content';
-export * from './dropdown-menu-item';
-export * from './dropdown-menu-label';
-export * from './dropdown-menu-radio-item';
-export * from './dropdown-menu-separator';
-export * from './dropdown-menu-shortcut';
-export * from './dropdown-menu-sub-content';
-export * from './dropdown-menu-sub-trigger';
-export * from './primitives';
