--- src/components/ui/dropdown-menu/primitives.ts
@@ -1,8 +1,0 @@
-export {
-  DropdownMenu,
-  DropdownMenuTrigger,
-  DropdownMenuGroup,
-  DropdownMenuPortal,
-  DropdownMenuSub,
-  DropdownMenuRadioGroup,
-} from '@radix-ui/react-dropdown-menu';
